% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addTrapRate.R
\name{addTrapRate}
\alias{addTrapRate}
\title{Combine fish and trap rate data}
\usage{
addTrapRate(
  lgr_weekly = NULL,
  trap_rate = NULL,
  trap_rate_dist = c("beta", "logit")
)
}
\arguments{
\item{lgr_weekly}{dataframe containing weekly summaries of window counts and trap data. Part of what is returned by \code{summariseLGRweekly}.}

\item{trap_rate}{dataframe containing estimates of trap rate}

\item{trap_rate_dist}{distributional form for trap rate prior. \code{beta} returns alpha and beta parameters for beta distribution. \code{logit} returns mean and standard deviation in logit space. Default is \code{beta}. Support for logit distribution is coming in future versions of STADEM.}
}
\description{
Combine weekly data with trap rate data, and determine which trap rate estimates are invalid
}
\author{
Kevin See
}
