% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chapmanMR.R
\name{chapmanMR}
\alias{chapmanMR}
\title{Modified Lincoln-Petersen (Chapman) estimator}
\usage{
chapmanMR(data, rmInvalid = TRUE)
}
\arguments{
\item{data}{data.frame containing columns labeled \code{M}, \code{C} and \code{R} for marks, captures and recaptures}

\item{rmInvalid}{if estimates are deemed invalid by Robson & Regier criteria, make them NA}
}
\description{
Estimate total abundance from a 2 pass mark - recapture event
}
\author{
Kevin See
}

