% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compileData.R
\name{compileData}
\alias{compileData}
\title{STADEM Data Summary}
\usage{
compileData(
  spp = c("Chinook", "Steelhead", "Coho"),
  spawn_yr = NULL,
  dam = c("LWG", "WFF", "BON", "TDA", "JDA", "MCN", "IHR", "LMN", "LGS", "PRO", "ROZ",
    "PRD", "WAN", "RIS", "TUM", "RRH", "WEL", "ZOS"),
  start_date = NULL,
  end_date = NULL,
  incl_jacks = NULL,
  sthd_type = c("all", "unclipped"),
  strata_beg = "Mon",
  last_strata_min = 3,
  sthd_B_run = FALSE,
  trap_dbase = NULL,
  incl_trapRate = T,
  useDARTrate = F,
  trap_rate_cv = 0,
  trap_rate_dist = c("beta", "logit")
)
}
\arguments{
\item{spawn_yr}{spawn year.}

\item{dam}{the dam code for the dam you wish to query for window counts. Possible codes are: WFF (Willamette Falls), BON (Bonneville), TDA (The Dalles), JDA (John Day), MCN (McNary), IHR (Ice Harbor), LMN (Lower Monumental), LGS (Little Goose), LWG (Lower Granite), PRO (Prosser), ROZ (Roza), PRD (Priest Rapids), WAN (Wanapum), RIS (Rock Island), TUM (Tumwater), RRH (Rocky Reach), WEL (Wells), ZOS (Zosel)}

\item{start_date}{character vector of date (\code{YYYYMMDD}) when query should start}

\item{end_date}{character vector of date (\code{YYYYMMDD}) when query should end}

\item{incl_jacks}{should jacks be included in the window count totals? \code{T / F} Default is `FALSE`}

\item{sthd_type}{should window counts of steelhead be for all steelhead, or only unclipped (i.e. wild) fish? Default is \code{all}.}

\item{strata_beg}{3 letter code for the day of the week each weekly strata
should begin on. Default value is \code{'Mon'}.}

\item{last_strata_min}{minimum length (in days) for the final strata. Default
value is 3.}

\item{sthd_B_run}{should numbers of B run steelhead be reported? These are
defined as wild steelhead greater than 780mm in length. Default is
\code{FALSE}.}

\item{trap_dbase}{data frame object containing the LWG trapping data with
identical data types as in tblLGDMasterCombineExport; required fields
include MasterID, LGDNumPIT, CollectionDate, SRR, LGDSpecies, LGDRear,
LGDLifeStage, LGDMarkAD, LGDValid, LGDFLmm, PTAgisSxCGRAObs.}

\item{useDARTrate}{should the DART query for the trap rate be used? Default
is \code{FALSE}, which implies the trap rate is estimated by PIT tags.}

\item{trap_rate_cv}{constant coefficient of variation (CV) that should be
applied to estimates of trap rate queried by DART. Default value is
\code{0}.}

\item{trap_rate_dist}{distributional form for trap rate prior. \code{beta}
returns alpha and beta parameters for beta distribution. \code{logit}
returns mean and standard deviation in logit space.}
}
\description{
Query and summarise data for STADEM
}
\examples{
compileData(2012)
}
\author{
Kevin See
}
