% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractPost.R
\name{extractPost}
\alias{extractPost}
\title{Extract STADEM Posteriors}
\usage{
extractPost(stadem_mod = NULL, param_nms = "X.new.wild")
}
\arguments{
\item{stadem_mod}{The result of `runSTADEMmodel()`. An MCMC.list, or \code{jagsUI} object that can be coerced to an MCMC.list.}

\item{param_nms}{Character vector with the name of the parameter corresponding to total escapement. Default is focused on wild fish, `X.new.wild`.}
}
\description{
Extracts the MCMC posteriors of selected parameters from a STADEM model, and formats them in a long format.
}
\examples{
extractPost()
}
\author{
Kevin See
}
