% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jagsInits.R
\name{jagsInits}
\alias{jagsInits}
\title{Initial values}
\usage{
jagsInits(jags_data)
}
\arguments{
\item{jags_data}{list of data being passed to \code{JAGS} model}
}
\description{
Function to set appropriate initial values
}
\author{
Kevin See
}

