% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrTrapRate.R
\name{mrTrapRate}
\alias{mrTrapRate}
\title{Mark-recapture trap rate}
\usage{
mrTrapRate(filepath = NULL, week_strata = NULL)
}
\arguments{
\item{filepath}{file path to the excel file containing data}

\item{week_strata}{vector of intervals delimiting the weekly strata to summarize mark-recapture data over. Returned as one part of the \code{summariseLGRweekly} function.}
}
\description{
Estimate the trap rate at Lower Granite by using a mark-recapture model. The trap is the marking event, and detections at the weir at the top of the fish ladder is the 2nd capture event.
}
\author{
Kevin See
}

