% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTimeSeries.R
\name{plotTimeSeries}
\alias{plotTimeSeries}
\title{Time Series Plot}
\usage{
plotTimeSeries(stadem_mod = NULL, weeklyData = NULL)
}
\arguments{
\item{stadem_mod}{jagsUI object returned from running \code{runSTADEMmodel}.}

\item{weeklyData}{weekly data compiled by \code{compileGRAdata}, with the name \code{weeklyData} in the resulting list.}
}
\description{
Plot window counts, trap estimates and STADEM estimates of total escapement through time.
}
\author{
Kevin See
}
