% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepJAGS.R
\name{prepJAGS}
\alias{prepJAGS}
\title{Compile JAGS data}
\usage{
prepJAGS(lgr_weekly = NULL, sthd_B_run = F, use_historical = F,
  type = c("PBT", "Morph"))
}
\arguments{
\item{lgr_weekly}{dataframe containing weekly summaries of window counts and trap data. Part of what is returned by \code{summariseLGRweekly}.}

\item{sthd_B_run}{should numbers of B run steelhead be incorporated? These are defined as wild steelhead greater than 780mm in length. Default is \code{FALSE}.}

\item{use_historical}{should older historical data about night passage rates be incorporated? Default is \code{FALSE}.}
}
\description{
Compile data into format for the JAGS total escapement model
}
\author{
Kevin See
}

