% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryLGRtrapDatabase.R
\name{queryLGRtrapDatabase}
\alias{queryLGRtrapDatabase}
\title{Access Lower Granite Dam trap database}
\usage{
queryLGRtrapDatabase(username = NULL, password = NULL)
}
\arguments{
\item{username}{A username to access the SharePoint site where the trap database resides.}

\item{password}{A password to access the SharePoint site where the trap database resides.}
}
\description{
Download copy of the Lower Granite Dam trap database. ** This currently doesn't work **
}
\examples{
# Don't run; # queryLGRtrapDatabase(username = 'ksee', password = 'IDFGuser')
}
\author{
Kevin See
}

