% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryPITtagData.R
\name{queryPITtagData}
\alias{queryPITtagData}
\title{PIT tag data}
\source{
\url{http://www.cbr.washington.edu/dart}
}
\usage{
queryPITtagData(dam = "GRA", spp_code = c("1", "2", "3", "4"), yr = NULL)
}
\arguments{
\item{dam}{the dam code for the dam you wish to query for PIT tag data. Currently only available for Lower Granite Dam (GRA).}

\item{spp_code}{species code to query window counts for. Possible codes are: 1 (Chinook), 2 (Coho), 3 (Steelhead), 4 (Sockeye)}

\item{yr}{calendar year to query for PIT tag data.}
}
\description{
Query and download PIT tag data about night passage and re-ascension at particular dams for specific species and year, using DART.
}
\examples{
queryPITtagData(yr = 2015)
}
\author{
Kevin See
}

