% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryPITtagObs.R
\name{queryPITtagObs}
\alias{queryPITtagObs}
\title{PIT tag data}
\source{
\url{http://www.cbr.washington.edu/dart}
}
\usage{
queryPITtagObs(site = "GRA", spp = c("Chinook", "Coho", "Steelhead",
  "Sockeye"), yr = NULL, start_day = NULL, end_day = NULL,
  span_yrs = NULL)
}
\arguments{
\item{site}{The site to query observations for.}

\item{spp}{species to query window counts for. Possible choices are: Chinook, Coho, Steelhead, Sockeye}

\item{yr}{year to query for window counts.}

\item{start_day}{date (\code{month / day}) when query should start}

\item{end_day}{date (\code{month / day}) when query should end}

\item{span_yrs}{Should the query span mulitple years? If \code{NULL}, default value is chosen based on species.}
}
\description{
Query and download data about PIT tags observed crossing a particular site, using DART.
}
\examples{
queryPITtagObs(yr = 2013)
}
\author{
Kevin See
}
