% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryTrapRate.R
\name{queryTrapRate}
\alias{queryTrapRate}
\title{DART reported trap rate}
\source{
\url{http://www.cbr.washington.edu/dart}
}
\usage{
queryTrapRate(week_strata = NULL, spp = c("Chinook", "Steelhead"),
  return_weekly = T)
}
\arguments{
\item{week_strata}{vector of intervals delimiting the weekly strata to summarize mark-recapture data over, as returned by \code{weeklyStrata()}.}

\item{spp}{species to query window counts for. Possible choices are: \code{Chinook} or \code{Steelhead}}

\item{return_weekly}{Should the results be summarised on weekly basis? Default is \code{TRUE}.}
}
\description{
Estimate the trap rate at Lower Granite by using a mark-recapture model. The trap is the marking event, and detections at the weir at the top of the fish ladder is the 2nd capture event.
}
\author{
Kevin See
}
