% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readLGRtrapDB.R
\name{readLGRtrapDB}
\alias{readLGRtrapDB}
\title{LGR Trap Database}
\usage{
readLGRtrapDB(filepath = "data-raw/tblLGDMasterCombineExportJodyW.csv",
  spring_summer_chnk = T)
}
\arguments{
\item{filepath}{path to the LGR trap database as csv file}

\item{spring_summer_chnk}{should the PIT tag data be filtered to exclude fall and winter run Chinook? Default is \code{TRUE}}
}
\description{
Read in data from LGR trap database
}
\examples{
# do not run; #readLGRtrapDB()
}
\author{
Kevin See
}

