% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summariseLGRweekly.R
\name{summariseLGRweekly}
\alias{summariseLGRweekly}
\title{Compile weekly data for Lower Granite dam}
\usage{
summariseLGRweekly(wind_data = NULL, pit_all = NULL, trap_df = NULL,
  incl_clip_sthd = FALSE, sthd_B_run = FALSE, incl_chnk_jacks = TRUE)
}
\arguments{
\item{wind_data}{data.frame containing window count data to summarise, queried using \code{queryWindCnts}, and potentially combined across calendar years}

\item{pit_all}{data.frame containing data about individual PIT tags detected at Lower Granite Dam, produced using \code{queryPITtagData}}

\item{trap_df}{data.frame as created by \code{readLGRtrapDB}.}

\item{incl_clip_sthd}{should clipped steelhead in the trap be included in this summary? Should match with the window counts. Default is \code{FALSE}.}

\item{incl_chnk_jacks}{should window counts of Chinook jacks be included with adult counts? Default is \code{TRUE}}
}
\description{
Combine window counts and PIT tag data on weekly time-steps
}
\author{
Kevin See
}
