% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarisePITdataDaily.R
\name{summarisePITdataDaily}
\alias{summarisePITdataDaily}
\title{Summarise PIT tag data}
\usage{
summarisePITdataDaily(pit_data)
}
\arguments{
\item{pit_data}{data.frame containing PIT tag data to summarise, queried using \code{queryPITtagData}}
}
\description{
Summarise night passage and re-ascension data from PIT tags at Lower Granite Dam on a daily time-step
}
\examples{
summarisePITdataDaily(pit_all)
}
\author{
Kevin See
}
