% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagTrapRate.R
\name{tagTrapRate}
\alias{tagTrapRate}
\title{PIT tag based trap rate}
\usage{
tagTrapRate(trap_dataframe = NULL, week_strata = NULL,
  mark_site = "LGRLDR")
}
\arguments{
\item{trap_dataframe}{dataframe containing data from the fish trap, including columns named \code{Tag.ID}, \code{CollectionDate} and \code{SRR}.}

\item{week_strata}{vector of intervals delimiting the weekly strata to summarize mark-recapture data over, as returned by the \code{weeklyStrata} function.}

\item{mark_site}{PTAGIS code for the site at the location being queried for trap rate. This is used to filter out fish that were marked at the fish trap.}
}
\description{
Estimate the trap rate at Lower Granite by using the proportion of previously PIT tagged fish observed pass a location that were found in the fish trap at that location.
}
\author{
Kevin See
}
