% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weeklyStrata.R
\name{weeklyStrata}
\alias{weeklyStrata}
\title{Generate weekly strata}
\usage{
weeklyStrata(
  start_date = NULL,
  end_date = NULL,
  strata_beg = "Mon",
  last_strata_min = 3
)
}
\arguments{
\item{start_date}{character vector of date (\code{YYYYMMDD}) when query should start}

\item{end_date}{character vector of date (\code{YYYYMMDD}) when query should end}

\item{strata_beg}{3 letter code for the day of the week each weekly strata should begin on. Default value is \code{'Mon'}.}

\item{last_strata_min}{minimum length (in days) for the final strata. Default value is 3.}
}
\description{
Based on spawn year and species
}
\examples{
weeklyStrata(2010:2012)
}
\author{
Kevin See
}
