% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addTrapRate.R
\name{addTrapRate}
\alias{addTrapRate}
\title{Combine fish and trap rate data}
\usage{
addTrapRate(lgr_weekly = NULL, trap_rate = NULL)
}
\arguments{
\item{lgr_weekly}{dataframe containing weekly summaries of window counts and trap data. Part of what is returned by \code{summariseLGRweekly}.}

\item{trap_rate}{dataframe containing estimates of trap rate}
}
\description{
Combine weekly data with trap rate data, and determine which trap rate estimates are invalid
}
\author{
Kevin See
}

