% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepJAGS.R
\name{prepJAGS}
\alias{prepJAGS}
\title{Compile JAGS data}
\usage{
prepJAGS(
  lgr_weekly = NULL,
  sthd_B_run = F,
  hw_type = c("PBT", "Morph"),
  wild_tags = F
)
}
\arguments{
\item{lgr_weekly}{dataframe containing weekly summaries of window counts and trap data. Part of what is returned by \code{summSTADEM}.}

\item{sthd_B_run}{should numbers of B run steelhead be incorporated? These are defined as wild steelhead greater than 780mm in length. Default is \code{FALSE}.}

\item{hw_type}{what criteria is used to determine hatchery vs. wild? \code{Morph} uses the morphilogical call made at the trap. \code{PBT} uses the genetic information gleaned from parental based tagging.}

\item{wild_tags}{Should only wild PIT tags be used to estimate daytime passage and re-ascension rates? Default is \code{FALSE}.}
}
\description{
Compile data into format for the JAGS total escapement model
}
\author{
Kevin See
}
