% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryPITtagData.R
\name{queryPITtagData}
\alias{queryPITtagData}
\title{PIT tag data}
\source{
\url{https://www.cbr.washington.edu/dart}
}
\usage{
queryPITtagData(
  damPIT = c("GRA", "PRA", "RIA", "RRF"),
  spp = c("Chinook", "Coho", "Steelhead", "Sockeye"),
  start_date = NULL,
  end_date = NULL
)
}
\arguments{
\item{damPIT}{the dam code for the dam you wish to query for PIT tag data. Currently only available for Lower Granite Dam (GRA), Priest Rapids (PRA), Rock Island (RIA) and Rocky Reach (RRF).}

\item{spp}{species to query PIT tag observations for. Possible choices are: `Chinook`, `Coho`, `Steelhead`, `Sockeye`}

\item{start_date}{character vector of date (\code{YYYYMMDD}) when query should start}

\item{end_date}{character vector of date (\code{YYYYMMDD}) when query should end}
}
\description{
Query and download PIT tag data about night passage and re-ascension at particular dams for specific species and year, using DART.
}
\examples{
queryPITtagData(spp = "Steelhead", start_date = "20180701")
}
\author{
Kevin See
}
