% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryPITtagObs.R
\name{queryPITtagObs}
\alias{queryPITtagObs}
\title{PIT tag data}
\source{
\url{https://www.cbr.washington.edu/dart}
}
\usage{
queryPITtagObs(
  site = "GRA",
  spp = c("Chinook", "Coho", "Steelhead", "Sockeye"),
  start_date = NULL,
  end_date = NULL
)
}
\arguments{
\item{site}{The site to query observations for.}

\item{spp}{species to query PIT tag observations for. Possible choices are: Chinook, Coho, Steelhead, Sockeye}

\item{start_date}{character vector of date (\code{YYYYMMDD}) when query should start}

\item{end_date}{character vector of date (\code{YYYYMMDD}) when query should end}
}
\description{
Query and download data about PIT tags observed crossing a particular site, using DART.
}
\examples{
queryPITtagObs(yr = 2013)
}
\author{
Kevin See
}
