% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryPITtagObs.R
\name{queryPITtagObs}
\alias{queryPITtagObs}
\title{PIT tag data}
\source{
\url{https://www.cbr.washington.edu/dart}
}
\usage{
queryPITtagObs(
  site = c("GRA", "PRA"),
  spp = c("Chinook", "Coho", "Steelhead", "Sockeye"),
  start_date = NULL,
  end_date = NULL
)
}
\arguments{
\item{site}{The site to query observations for. Query only set up for Lower Granite Dam Adult Fishway (GRA) and Priest Rapids fishway (PRA) at this time.}

\item{spp}{species to query PIT tag observations for. Possible choices are: `Chinook`, `Coho`, `Steelhead`, `Sockeye`}

\item{start_date}{character vector of date (\code{YYYYMMDD}) when query should start}

\item{end_date}{character vector of date (\code{YYYYMMDD}) when query should end}
}
\description{
Query and download data about PIT tags observed crossing a particular site, using DART.
}
\examples{
queryPITtagObs(spp = "Steelhead", start_date = "20180701")
}
\author{
Kevin See
}
