% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryTrapRate.R
\name{queryTrapRate}
\alias{queryTrapRate}
\title{DART reported trap rate}
\source{
\url{https://www.cbr.washington.edu/dart}
}
\usage{
queryTrapRate(
  week_strata = NULL,
  spp = c("Steelhead", "Chinook"),
  return_weekly = T
)
}
\arguments{
\item{week_strata}{vector of intervals delimiting the weekly strata to summarize mark-recapture data over, as returned by \code{weeklyStrata()}.}

\item{spp}{species to query trap rate for. Possible choices are: \code{Chinook} or \code{Steelhead}}

\item{return_weekly}{Should the results be summarised on weekly basis? Default is \code{TRUE}.}
}
\description{
Query DART for the reported trap rate at Lower Granite dam.
}
\author{
Kevin See
}
