% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readLGRtrapDB.R
\name{readLGRtrapDB}
\alias{readLGRtrapDB}
\title{LGR Trap Database}
\usage{
readLGRtrapDB(trap_path = NULL, date_range = NULL)
}
\arguments{
\item{trap_path}{file path where a csv file containing the data from the fish trap is located}

\item{date_range}{vector of length 2, with minimum and maximum dates to read.}
}
\description{
Read in data from LGR trap database
}
\examples{
# do not run; #readLGRtrapDB()
}
\author{
Kevin See
}
