% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readLGRtrapDB.R
\name{readLGRtrapDB}
\alias{readLGRtrapDB}
\title{LGR Trap Database}
\usage{
readLGRtrapDB(filepath = "data-raw/tblLGDMasterCombineExportJodyW.csv",
  date_range = NULL)
}
\arguments{
\item{filepath}{path to the LGR trap database as csv file}

\item{date_range}{vector of length 2, with minimum and maximum dates to read.}
}
\description{
Read in data from LGR trap database
}
\examples{
# do not run; #readLGRtrapDB()
}
\author{
Kevin See
}
