% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trapRateInputs.R
\name{trapRateInputs}
\alias{trapRateInputs}
\title{Combined trap rate}
\source{
\url{http://www.cbr.washington.edu/dart}
}
\usage{
trapRateInputs(filepath = NULL, trap_dataframe = NULL, week_strata = NULL,
  return_weekly = T, poor_cv_threshold = 0.7, m = c("Mt.bc", "Mt", "Chap",
  NULL))
}
\arguments{
\item{filepath}{file path to the excel file containing data}

\item{week_strata}{vector of intervals delimiting the weekly strata to summarize mark-recapture data over. Returned as one part of the \code{summariseLGRweekly} function.}

\item{return_weekly}{Should the results be summarised on weekly basis? Default is \code{TRUE}.}

\item{poor_cv_threshold}{The threshold for the CV of the mark-recapture estimate of trap rate. If an estimate exceeds this threshold, the calculated rate from DART will be used.}

\item{m}{model type to use to estimate parameters. Standard \code{Mt} and \code{Mt.bc} use \code{closedp} and \code{closedp.bc} functions from \code{Rcapture} package. \code{Chap} used the Chapman, or modified Lincoln-Peterson estimator.}
}
\description{
Utilize mark-recapture estimates of trap rate, as well as observed rates queried through DART.
}
\author{
Kevin See
}
