% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weeklyStrata.R
\name{weeklyStrata}
\alias{weeklyStrata}
\title{Generate weekly strata}
\usage{
weeklyStrata(spawn_yr, spp = c("Chinook", "Steelhead"), start_day = NULL,
  end_day = NULL, strata_beg = NULL, last_strata_min = 3)
}
\arguments{
\item{spawn_yr}{spawn year to divide into weekly strata}

\item{spp}{choices are either \code{Chinook} or \code{Steelhead}}

\item{start_day}{date (\code{month / day}) when strata should start}

\item{end_day}{date (\code{month / day}) when strata should end}

\item{strata_beg}{3 letter code for the day of the week each weekly strata should begin on.}

\item{last_strata_min}{minimum length (in days) for the final strata}
}
\description{
Based on spawn year and species
}
\examples{
weeklyStrata(2010:2012)
}
\author{
Kevin See
}
