% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weeklyStrata.R
\name{weeklyStrata}
\alias{weeklyStrata}
\title{Generate weekly strata}
\usage{
weeklyStrata(years, start_day = c("0701", "0301"), end_day = NULL)
}
\arguments{
\item{years}{which years should go into the weekly strata}

\item{start_day}{day to start strata, in the format \code{month} \code{day}}

\item{end_day}{day to end strata, if needed}
}
\description{
Based on year range, and start dates
}
\examples{
weeklyStrata(2010:2012)
}
\author{
Kevin See
}

