% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeJAGSmodel.R
\name{writeJAGSmodel}
\alias{writeJAGSmodel}
\title{Write JAGS model}
\usage{
writeJAGSmodel(file_name = NULL, use_historical = F)
}
\arguments{
\item{file_name}{name (with file path) to save the model as}

\item{use_historical}{should older historical data about night passage rates be incorporated? Default is \code{FALSE}.}
}
\description{
Writes a text file version of the model to be used by JAGS
}
\examples{
writeJAGSmodel()


}
\author{
Kevin See
}

